+++ lib/searchkick/reindex.rb
@@ -7,7 +7,7 @@ module Searchkick
       skip_import = options[:import] == false
 
       alias_name = searchkick_index.name
-      new_name = alias_name + "_" + Time.now.strftime("%Y%m%d%H%M%S%L")
+      new_name = "#{alias_name}_#{Time.now.strftime('%Y%m%d%H%M%S%L')}"
       index = Searchkick::Index.new(new_name)
 
       clean_indices
@@ -40,7 +40,7 @@ module Searchkick
     # remove old indices that start w/ index_name
     def clean_indices
       all_indices = Searchkick.client.indices.get_aliases
-      indices = all_indices.select{|k, v| v["aliases"].empty? && k =~ /\A#{Regexp.escape(searchkick_index.name)}_\d{14,17}\z/ }.keys
+      indices = all_indices.select{|k, v| (v.empty? || v["aliases"].empty?) && k =~ /\A#{Regexp.escape(searchkick_index.name)}_\d{14,17}\z/ }.keys
       indices.each do |index|
         Searchkick::Index.new(index).delete
       end
@@ -203,7 +203,7 @@ module Searchkick
           }
         }
 
-        if searchkick_env == "test"
+        if Searchkick.env == "test"
           settings.merge!(number_of_shards: 1, number_of_replicas: 0)
         end
 
@@ -288,9 +288,16 @@ module Searchkick
           mapping[field] = field_mapping
         end
 
-        (options[:locations] || []).each do |field|
+        (options[:locations] || []).map(&:to_s).each do |field|
           mapping[field] = {
             type: "geo_point"
+          }
+        end
+
+        (options[:unsearchable] || []).map(&:to_s).each do |field|
+          mapping[field] = {
+            type: "string",
+            index: "no"
           }
         end