+++ test/facets_test.rb
@@ -1,4 +1,5 @@
 require_relative "test_helper"
+require "active_support/core_ext"
 
 class TestFacets < Minitest::Test