+++ test/test_helper.rb
@@ -53,6 +53,7 @@ if defined?(Mongoid)
     field :color
     field :latitude, type: BigDecimal
     field :longitude, type: BigDecimal
+    field :description
   end
 
   class Store
@@ -65,6 +66,45 @@ if defined?(Mongoid)
     include Mongoid::Document
 
     field :name
+  end
+
+  class Dog < Animal
+  end
+
+  class Cat < Animal
+  end
+elsif defined?(NoBrainer)
+  NoBrainer.configure do |config|
+    config.app_name = :searchkick
+    config.environment = :test
+  end
+
+  class Product
+    include NoBrainer::Document
+    include NoBrainer::Document::Timestamps
+
+    field :name,         type: String
+    field :store_id,     type: Integer
+    field :in_stock,     type: Boolean
+    field :backordered,  type: Boolean
+    field :orders_count, type: Integer
+    field :price,        type: Integer
+    field :color,        type: String
+    field :latitude
+    field :longitude
+    field :description,  type: String
+  end
+
+  class Store
+    include NoBrainer::Document
+
+    field :name, type: String
+  end
+
+  class Animal
+    include NoBrainer::Document
+
+    field :name, type: String
   end
 
   class Dog < Animal
@@ -95,6 +135,7 @@ else
     t.string :color
     t.decimal :latitude, precision: 10, scale: 7
     t.decimal :longitude, precision: 10, scale: 7
+    t.text :description
     t.timestamps
   end
 
@@ -146,7 +187,8 @@ class Product
     word_start: [:name],
     word_middle: [:name],
     word_end: [:name],
-    highlight: [:name]
+    highlight: [:name],
+    unsearchable: [:description]
 
   attr_accessor :conversions, :user_ids, :aisle